requirements.txt
anyio==4.12.1
attrs==25.4.0
boto3==1.42.30
botocore==1.42.30
CacheControl==0.14.3
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chembl-webresource-client==0.10.9
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
easydict==1.13
exceptiongroup==1.3.1
filelock==3.19.1
firebase_admin==7.1.0
flask==3.0.0
flask-cors==4.0.0
fonttools==4.60.1
fsspec==2025.10.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rdkit==2025.9.1
requests==2.32.5
requests-cache==1.2.1
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==1.26.20
wordcloud==1.9.4
zipp==3.23.0
